fix(playback): clear Auto artwork detekt/Sonar complexity and duplication
Extract shared AutoArtworkDownloader and AutoCollageLayouts, and split
prewarm helpers so CI detekt and Sonar cognitive-complexity/duplication gates pass.
