+++ core/playback/README.md
@@ -22,7 +22,8 @@ Owns playback session control, queue orchestration, smart queue logic, Media3 pl
 - Android Auto browse artwork:
   - `AutoArtworkRepository` + `AutoArtworkSourceStore` register remote/local sources into an in-memory map and durable `android_auto_artwork_sources` prefs (`commit`, not async `apply`) before returning `content://…/art|local|collage/…` URIs.
   - `AutoCollageProvider` lazily fetches remote covers with validated HTTPS redirects, lenient image content-types, magic-byte checks, and one retry; folder collage URIs include a `v=` cache-buster so Auto hosts reload when resume/history content changes.
-  - `AutoCollagePrewarmer` / `AutoCollageGenerator` rebuild section collages from content keys (resume episode IDs, queue IDs, subscriptions, …), use a shorter TTL for partial/fallback tiles, and refresh on mark-complete / queue changes via `AutoBrowseLibraryHost.requestAutoCollageRefresh`.
+  - `AutoCollagePrewarmer` / `AutoCollageGenerator` / `AutoCollageLayouts` rebuild section collages from content keys (resume episode IDs, queue IDs, subscriptions, …), use a shorter TTL for partial/fallback tiles, and refresh on mark-complete / queue changes via `AutoBrowseLibraryHost.requestAutoCollageRefresh`.
+  - `AutoArtworkDownloader` is the shared HTTPS fetch path (validated redirects + public-host checks) used by both collage generation and the ContentProvider.
 
 ## Internal structure
