+++ core/playback/src/main/java/cx/aswin/boxlore/core/playback/service/AutoCollagePrewarmer.kt
@@ -5,7 +5,10 @@ import android.net.Uri
 import android.util.Log
 import androidx.media3.session.MediaLibraryService.MediaLibrarySession
 import cx.aswin.boxlore.core.database.BoxLoreDatabase
+import cx.aswin.boxlore.core.database.DownloadedEpisodeEntity
+import cx.aswin.boxlore.core.database.ListeningHistoryEntity
 import cx.aswin.boxlore.core.database.PodcastEntity
+import cx.aswin.boxlore.core.model.Episode
 import cx.aswin.boxlore.core.model.Podcast
 import cx.aswin.boxlore.core.playback.MixtapeEngine
 import cx.aswin.boxlore.core.playback.QueueRepository
@@ -51,137 +54,154 @@ internal class AutoCollagePrewarmer(
 
     private suspend fun runPrewarm() {
         try {
-            val history = database.listeningHistoryDao().getRecentHistoryList(300)
-            val resumeItems = database.listeningHistoryDao().getResumeItemsList()
-            val subscriptions = database.podcastDao().getSubscribedPodcastsList()
-            val downloads = database.downloadedEpisodeDao().getCompletedDownloads(8)
-            val queue = queueRepository.getQueueSnapshot()
-            val historyImages =
-                history.mapNotNull {
-                    it.episodeImageUrl ?: it.podcastImageUrl
-                }
-            val resumeImages =
-                resumeItems.mapNotNull {
-                    it.episodeImageUrl ?: it.podcastImageUrl
-                }
-            val subscriptionImages = subscriptions.mapNotNull { it.imageUrl }
-            val downloadImages =
-                downloads.mapNotNull {
-                    it.episodeImageUrl ?: it.podcastImageUrl
-                }
-            val queueImages = queue.mapNotNull { it.imageUrl ?: it.podcastImageUrl }
-            val newEpisodeImages =
-                subscriptions.mapNotNull {
-                    it.latestEpisode?.imageUrl ?: it.imageUrl
-                }
-            var mixtape =
-                MixtapeEngine.build(
-                    subscriptions = subscriptions.map(toAutoPodcast),
-                    history = history,
-                    adaptiveRanking =
-                        MixtapeEngine.AdaptiveRanking(
-                            scorer = adaptiveCandidateScorer,
-                            surface = RankingSurface.ANDROID_AUTO,
-                        ),
+            val snapshot = loadSnapshot()
+            val mixtape = resolveMixtape(snapshot)
+            val uris =
+                AutoCollageGenerator.generateAllCollages(
+                    context = context,
+                    folderImages = snapshot.folderImages(mixtape),
+                    folderContentKeys = snapshot.folderContentKeys(mixtape),
                 )
-            if (mixtape.episodes.size < 3) {
-                val recommendations =
-                    runCatching {
-                        withTimeout(6_000L) {
-                            smartQueueSources.getPersonalizedRecommendations(
-                                history = smartQueueSources.getHistoryForRecommendations(25),
-                                interests = smartQueueSources.getInterests(),
-                                country = smartQueueSources.getRegion(),
-                                subscribedPodcastIds = subscriptions.map { it.podcastId },
-                                subscribedGenres = subscriptions.mapNotNull { it.genre }.distinct(),
-                            )
-                        }
-                    }.getOrDefault(emptyList())
-                mixtape =
-                    MixtapeEngine.build(
-                        subscriptions = subscriptions.map(toAutoPodcast),
-                        history = history,
-                        recommendations = recommendations,
-                        adaptiveRanking =
-                            MixtapeEngine.AdaptiveRanking(
-                                scorer = adaptiveCandidateScorer,
-                                surface = RankingSurface.ANDROID_AUTO,
-                            ),
+            onCollagesReady(uris)
+            notifyBrowseTree(snapshot.subscriptions.size, snapshot.resumeItems.size)
+        } catch (error: Exception) {
+            Log.w("AutoBrowse", "Unable to prewarm Android Auto artwork", error)
+        }
+    }
+
+    private suspend fun loadSnapshot(): PrewarmSnapshot {
+        val history = database.listeningHistoryDao().getRecentHistoryList(300)
+        val resumeItems = database.listeningHistoryDao().getResumeItemsList()
+        val subscriptions = database.podcastDao().getSubscribedPodcastsList()
+        val downloads = database.downloadedEpisodeDao().getCompletedDownloads(8)
+        val queue = queueRepository.getQueueSnapshot()
+        return PrewarmSnapshot(
+            history = history,
+            resumeItems = resumeItems,
+            subscriptions = subscriptions,
+            downloads = downloads,
+            queue = queue,
+        )
+    }
+
+    private suspend fun resolveMixtape(snapshot: PrewarmSnapshot): MixtapeEngine.Result {
+        var mixtape =
+            MixtapeEngine.build(
+                subscriptions = snapshot.subscriptions.map(toAutoPodcast),
+                history = snapshot.history,
+                adaptiveRanking =
+                    MixtapeEngine.AdaptiveRanking(
+                        scorer = adaptiveCandidateScorer,
+                        surface = RankingSurface.ANDROID_AUTO,
+                    ),
+            )
+        if (mixtape.episodes.size >= 3) return mixtape
+        val recommendations =
+            runCatching {
+                withTimeout(6_000L) {
+                    smartQueueSources.getPersonalizedRecommendations(
+                        history = smartQueueSources.getHistoryForRecommendations(25),
+                        interests = smartQueueSources.getInterests(),
+                        country = smartQueueSources.getRegion(),
+                        subscribedPodcastIds = snapshot.subscriptions.map { it.podcastId },
+                        subscribedGenres =
+                            snapshot.subscriptions.mapNotNull { it.genre }.distinct(),
                     )
+                }
+            }.getOrDefault(emptyList())
+        return MixtapeEngine.build(
+            subscriptions = snapshot.subscriptions.map(toAutoPodcast),
+            history = snapshot.history,
+            recommendations = recommendations,
+            adaptiveRanking =
+                MixtapeEngine.AdaptiveRanking(
+                    scorer = adaptiveCandidateScorer,
+                    surface = RankingSurface.ANDROID_AUTO,
+                ),
+        )
+    }
+
+    private fun notifyBrowseTree(
+        subscriptionCount: Int,
+        resumeCount: Int,
+    ) {
+        val session = mediaSessionProvider()
+        session?.notifyChildrenChanged(AutoBrowseContract.ROOT_ID, 4, null)
+        session?.notifyChildrenChanged(AutoBrowseContract.HOME_ID, 3, null)
+        session?.notifyChildrenChanged(
+            AutoBrowseContract.LIBRARY_ID,
+            subscriptionCount + 2,
+            null,
+        )
+        session?.notifyChildrenChanged(AutoBrowseContract.DISCOVER_ID, 3, null)
+        session?.notifyChildrenChanged(AutoBrowseContract.HOME_CONTINUE_ID, resumeCount, null)
+    }
+
+    private data class PrewarmSnapshot(
+        val history: List<ListeningHistoryEntity>,
+        val resumeItems: List<ListeningHistoryEntity>,
+        val subscriptions: List<PodcastEntity>,
+        val downloads: List<DownloadedEpisodeEntity>,
+        val queue: List<Episode>,
+    ) {
+        private val historyImages =
+            history.mapNotNull { it.episodeImageUrl ?: it.podcastImageUrl }
+        private val resumeImages =
+            resumeItems.mapNotNull { it.episodeImageUrl ?: it.podcastImageUrl }
+        private val subscriptionImages = subscriptions.mapNotNull { it.imageUrl }
+        private val downloadImages =
+            downloads.mapNotNull { it.episodeImageUrl ?: it.podcastImageUrl }
+        private val queueImages = queue.mapNotNull { it.imageUrl ?: it.podcastImageUrl }
+        private val newEpisodeImages =
+            subscriptions.mapNotNull { it.latestEpisode?.imageUrl ?: it.imageUrl }
+        private val newEpisodeKeys =
+            subscriptions.mapNotNull { podcast ->
+                podcast.latestEpisode?.id ?: podcast.podcastId.takeIf {
+                    podcast.latestEpisode != null || !podcast.imageUrl.isNullOrBlank()
+                }
             }
+
+        fun folderImages(mixtape: MixtapeEngine.Result): Map<String, List<String>> {
             val mixtapeImages =
                 mixtape.podcasts.mapNotNull { podcast ->
                     podcast.latestEpisode?.let { episode ->
                         episode.imageUrl ?: episode.podcastImageUrl ?: podcast.imageUrl
                     }
                 }
-            val newEpisodeKeys =
-                subscriptions.mapNotNull { podcast ->
-                    podcast.latestEpisode?.id ?: podcast.podcastId.takeIf {
-                        podcast.latestEpisode != null || !podcast.imageUrl.isNullOrBlank()
-                    }
-                }
-            val uris =
-                AutoCollageGenerator.generateAllCollages(
-                    context = context,
-                    folderImages =
-                        mapOf(
-                            AutoBrowseContract.HOME_ID to (historyImages + newEpisodeImages).take(4),
-                            AutoBrowseContract.LIBRARY_ID to subscriptionImages.take(4),
-                            AutoBrowseContract.DOWNLOADS_ID to downloadImages.take(4),
-                            AutoBrowseContract.DISCOVER_ID to subscriptionImages.asReversed().take(4),
-                            AutoBrowseContract.HOME_CONTINUE_ID to resumeImages.take(4),
-                            AutoBrowseContract.HOME_QUEUE_ID to queueImages.take(4),
-                            AutoBrowseContract.HOME_NEW_EPISODES_ID to newEpisodeImages.take(4),
-                            AutoBrowseContract.HOME_DRIVE_MIX_ID to mixtapeImages.take(4),
-                            AutoBrowseContract.DISCOVER_DRIVE_PICKS_ID to
-                                (queueImages + subscriptionImages).take(4),
-                            AutoBrowseContract.DISCOVER_TIME_PICKS_ID to emptyList(),
-                            AutoBrowseContract.DISCOVER_GENRES_ID to emptyList(),
-                        ),
-                    folderContentKeys =
-                        mapOf(
-                            AutoBrowseContract.HOME_ID to
-                                history.take(4).map { it.episodeId },
-                            AutoBrowseContract.LIBRARY_ID to
-                                subscriptions.take(4).map { it.podcastId },
-                            AutoBrowseContract.DOWNLOADS_ID to
-                                downloads.map { it.episodeId },
-                            AutoBrowseContract.DISCOVER_ID to
-                                subscriptions.asReversed().take(4).map { it.podcastId },
-                            AutoBrowseContract.HOME_CONTINUE_ID to
-                                resumeItems.map { it.episodeId },
-                            AutoBrowseContract.HOME_QUEUE_ID to
-                                queue.map { it.id },
-                            AutoBrowseContract.HOME_NEW_EPISODES_ID to
-                                newEpisodeKeys.take(4),
-                            AutoBrowseContract.HOME_DRIVE_MIX_ID to
-                                mixtape.episodes.map { it.id },
-                            AutoBrowseContract.DISCOVER_DRIVE_PICKS_ID to
-                                (
-                                    queue.map { it.id } +
-                                        subscriptions.map { it.podcastId }
-                                ).take(4),
-                            AutoBrowseContract.DISCOVER_TIME_PICKS_ID to
-                                listOf(AutoBrowseContract.DISCOVER_TIME_PICKS_ID),
-                            AutoBrowseContract.DISCOVER_GENRES_ID to
-                                listOf(AutoBrowseContract.DISCOVER_GENRES_ID),
-                        ),
-                )
-            onCollagesReady(uris)
-            val session = mediaSessionProvider()
-            session?.notifyChildrenChanged(AutoBrowseContract.ROOT_ID, 4, null)
-            session?.notifyChildrenChanged(AutoBrowseContract.HOME_ID, 3, null)
-            session?.notifyChildrenChanged(
-                AutoBrowseContract.LIBRARY_ID,
-                subscriptions.size + 2,
-                null,
+            return mapOf(
+                AutoBrowseContract.HOME_ID to (historyImages + newEpisodeImages).take(4),
+                AutoBrowseContract.LIBRARY_ID to subscriptionImages.take(4),
+                AutoBrowseContract.DOWNLOADS_ID to downloadImages.take(4),
+                AutoBrowseContract.DISCOVER_ID to subscriptionImages.asReversed().take(4),
+                AutoBrowseContract.HOME_CONTINUE_ID to resumeImages.take(4),
+                AutoBrowseContract.HOME_QUEUE_ID to queueImages.take(4),
+                AutoBrowseContract.HOME_NEW_EPISODES_ID to newEpisodeImages.take(4),
+                AutoBrowseContract.HOME_DRIVE_MIX_ID to mixtapeImages.take(4),
+                AutoBrowseContract.DISCOVER_DRIVE_PICKS_ID to
+                    (queueImages + subscriptionImages).take(4),
+                AutoBrowseContract.DISCOVER_TIME_PICKS_ID to emptyList(),
+                AutoBrowseContract.DISCOVER_GENRES_ID to emptyList(),
             )
-            session?.notifyChildrenChanged(AutoBrowseContract.DISCOVER_ID, 3, null)
-            session?.notifyChildrenChanged(AutoBrowseContract.HOME_CONTINUE_ID, resumeItems.size, null)
-        } catch (error: Exception) {
-            Log.w("AutoBrowse", "Unable to prewarm Android Auto artwork", error)
         }
+
+        fun folderContentKeys(mixtape: MixtapeEngine.Result): Map<String, List<String>> =
+            mapOf(
+                AutoBrowseContract.HOME_ID to history.take(4).map { it.episodeId },
+                AutoBrowseContract.LIBRARY_ID to subscriptions.take(4).map { it.podcastId },
+                AutoBrowseContract.DOWNLOADS_ID to downloads.map { it.episodeId },
+                AutoBrowseContract.DISCOVER_ID to
+                    subscriptions.asReversed().take(4).map { it.podcastId },
+                AutoBrowseContract.HOME_CONTINUE_ID to resumeItems.map { it.episodeId },
+                AutoBrowseContract.HOME_QUEUE_ID to queue.map { it.id },
+                AutoBrowseContract.HOME_NEW_EPISODES_ID to newEpisodeKeys.take(4),
+                AutoBrowseContract.HOME_DRIVE_MIX_ID to mixtape.episodes.map { it.id },
+                AutoBrowseContract.DISCOVER_DRIVE_PICKS_ID to
+                    (queue.map { it.id } + subscriptions.map { it.podcastId }).take(4),
+                AutoBrowseContract.DISCOVER_TIME_PICKS_ID to
+                    listOf(AutoBrowseContract.DISCOVER_TIME_PICKS_ID),
+                AutoBrowseContract.DISCOVER_GENRES_ID to
+                    listOf(AutoBrowseContract.DISCOVER_GENRES_ID),
+            )
     }
 
     companion object {
